+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to NetVips will be documented in this file. See [here](CHANG
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [3.0.1] - TBD
+### Changed
+- Increase severity of native library load failures.
+
 ### Fixed
 - Load the correct library on all Apple platforms ([#240](https://github.com/kleisauke/net-vips/issues/240)).